🔨 Refactor import paths for grapher module
