+++ apps/backport/backport.py
@@ -21,7 +21,7 @@ from etl import grapher_model as gm
 from etl.backport_helpers import GrapherConfig
 from etl.db import get_engine, read_sql
 from etl.files import checksum_str
-from etl.grapher_io import variable_data_df_from_s3
+from etl.grapher.io import variable_data_df_from_s3
 from etl.snapshot import Snapshot, SnapshotMeta
 
 from . import utils
