+++ apps/wizard/app_pages/indicator_upgrade/indicator_mapping.py
@@ -17,7 +17,7 @@ from apps.wizard.utils import set_states
 from apps.wizard.utils.components import Pagination
 from etl.config import OWID_ENV
 from etl.db import get_engine, read_sql
-from etl.grapher_io import variable_data_df_from_s3
+from etl.grapher.io import variable_data_df_from_s3
 
 # Logger
 log = get_logger()
