+++ apps/wizard/app_pages/indicator_upgrade/utils.py
@@ -10,7 +10,7 @@ from structlog import get_logger
 
 from apps.wizard.utils.io import get_steps_df
 from etl.db import get_connection
-from etl.grapher_io import get_all_datasets, get_dataset_charts, get_variables_in_dataset
+from etl.grapher.io import get_all_datasets, get_dataset_charts, get_variables_in_dataset
 from etl.match_variables import find_mapping_suggestions, preliminary_mapping
 
 # Logger
