+++ apps/wizard/utils/cached.py
@@ -11,7 +11,7 @@ from sqlalchemy.orm import Session
 
 import etl.grapher_model as gm
 from apps.utils.map_datasets import get_grapher_changes
-from etl import grapher_io as gio
+from etl.grapher import io as gio
 from etl.config import OWID_ENV, OWIDEnv
 from etl.db import get_engine
 from etl.git_helpers import get_changed_files
