+++ apps/wizard/utils/chart_config.py
@@ -6,7 +6,7 @@ from typing import Any, Dict, List, Optional
 import numpy as np
 
 from etl.config import OWID_ENV, OWIDEnv
-from etl.grapher_io import ensure_load_variable
+from etl.grapher.io import ensure_load_variable
 from etl.grapher_model import Variable
 
 CONFIG_BASE = {
