+++ etl/scripts/faostat/archive/migrate_to_new_metadata.py
@@ -13,7 +13,7 @@ from owid import catalog
 from structlog import get_logger
 
 from etl.files import yaml_dump
-from etl.grapher_io import get_dataset_id, get_variables_in_dataset
+from etl.grapher.io import get_dataset_id, get_variables_in_dataset
 from etl.paths import DATA_DIR, STEP_DIR
 
 # Initialize logger.
